Idiotproof usage: show a basic example (#61)
Rationale:

1. Some don't remember dd incanatation and how many bytes are needed
   exactly for a key.

2. base64 is bloody complicated.

3. Even when confronted with the usage screen - it isn't immediately
   clear what is the quickest way to start the vpn; even now I still
   refer to the project README anyway.

base64 usage is hard to remember because it's inconsistent across
platforms -- but can be used consistently. It's partly the
luck of the draw depending on which platforms you interact with.

First it is hard to specify input for base64 if you try too hard.
Most implementations thankfully use stdin by default for encode
and decode.

Remember the encode flag is hard. There's -e and --encode. Except
as far as I can tell it's only for FreeBSD. Linux and Mac both
default to encode without even providing an option.

Ditto for decoding. It's hard to remember if it is:
base64 -d or -D or --decode. It is -d on Linux/FreeBSD, -D on mac,
but fortunately --decode across all.

This usage screen puts an end to the madness and gives the most common
ways to do it across platforms, which may not work 99% of the time.
Otherwise, I'm unnecessarily spending a few minutes to
github.com/jedisct1/dsvpn and ctrl-f for dd/base64.

And frankly the usage screen is probably biggest selling point of the
program. Even despite how trivial it is to deploy OpenVPN with a
single config file + a few certificates :).