Makefile: install command installs into sbin/ directory
The install command at the moment installs to "sbin" without
instead of "sbin/". This frustrates packaging because in the
former a file can be created whereas in the latter if the
directory doesn't exist make will fail.

The latter is of course the desirable choice. Nobody wants
to install a dsvpn binary called sbin.